cleanup lib/colors
