+++ frontend/src/metabase/lib/colors/colors.ts
@@ -5,7 +5,7 @@ import type { ColorSettings } from "metabase-types/api/settings";
 
 import { METABASE_DARK_THEME } from "./constants/dark";
 import { METABASE_LIGHT_THEME } from "./constants/light";
-import type { MetabaseColorKey } from "./types";
+import type { MetabaseColorKey } from "./types/color-keys";
 
 const win = typeof window !== "undefined" ? window : ({} as Window);
 const tokenFeatures = win.MetabaseBootstrap?.["token-features"] ?? {};
