+++ frontend/src/metabase/lib/colors/index.ts
@@ -2,14 +2,9 @@ export * from "./palette";
 export * from "./colors";
 export * from "./theme";
 
-// Re-export from submodules for convenience
-export { PROTECTED_COLORS } from "./constants/protected-colors";
 export { METABASE_DARK_THEME } from "./constants/dark";
 export { METABASE_LIGHT_THEME } from "./constants/light";
-export type {
-  ChartColorV2,
-  MetabaseColorKey,
-  MetabaseThemeV2,
-  UserThemeOverride,
-  ProtectedColorKey,
-} from "./types";
+export { PROTECTED_COLORS } from "./constants/protected-colors";
+
+export type { ChartColorV2, MetabaseThemeV2, ProtectedColorKey } from "./types";
+export type { MetabaseColorKey } from "./types/color-keys";
